Change type for uPortGetTickTimeMs() to int32_t
ubxlib uses int32_t and wrap arounds will not matter for us.

+++ inc/u_cx_at_config.h
@@ -40,7 +40,7 @@
 
 /* Porting layer for getting time in ms.*/
 #ifndef U_CX_PORT_GET_TIME_MS
-extern uint64_t uPortGetTickTimeMs(void);
+extern int32_t uPortGetTickTimeMs(void);
 # define U_CX_PORT_GET_TIME_MS()   uPortGetTickTimeMs()
 #endif
